feat!: next12, local docs, md/mdx support (#13)
This has been a few months in the works and there may be a few new issues here and there, but overall we're very happy to finally release the next major version of the Next Product Docs for Next.js.

The biggest change is the support for "local" docs, in the same repository as the website and/or product itself. This was a requirement for the [Cert-Manager](https://cert-manager.io) docs, where the website and docs are coupled. It also allows a range of combinations. We added support to skip the `root path` to support pure `docs` repositories.

Another breaking change is the move from environment variables to options/props. This was getting a bit out of hand, so we switched to a set of options to configure both the path and props functions. Please see the readme for further information.

We've done our best to test this with previous docs versions, if we missed anything please [open an issue](https://github.com/zentered/next-product-docs/issues/new) and describe the problem.

* feat: use local docs folder
* refactor: esm and improved re-usability for other projects
* fix: overhaul relative links lib and tests, fix variable scope
* fix: relative links with target and trailing slash
* fix: local docs adjustments
* feat: add pnpm link steps
* fix: remove scroll-element and prefix
* fix!: next12 dynamic routing https://nextjs.org/docs/routing/dynamic-routes\#catch-all-routes
* feat!: use options instead of environment variables
* feat: re-add sections
* feat: re-add tabs
* fix: serialization + tests
* chore: update dependencies
* chore: readme adjustments

Co-authored-by: Mo Sattler <hi@mosattler.com>

BREAKING CHANGE: the entire library has been refactored to ESM and updated to the latest libraries. Please update with caution.